backend/requirements.txt
# Web framework
django>=5.0,<6.0
djangorestframework>=3.15
django-cors-headers>=4.3

# MongoDB ODM
mongoengine>=0.28
pymongo>=4.7

# AI / LangChain stack
langchain>=0.3
langchain-community>=0.3
langchain-ollama>=0.2
langchain-chroma>=0.1
chromadb>=0.5

# Embeddings model (sentence-transformers)
sentence-transformers>=3.0

# Utilities
python-dotenv>=1.0
